requirement.txt
django==3.1.4
django-ckeditor==6.0.0